+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation: Docker Compose guide (EN/RU)
 - Documentation: expanded variable substitution docs in sendry-web guide (EN/RU)
 
+### Fixed
+- Sendry Web: template deploy now converts {{var}} to {{.var}} for Go template compatibility
+
 ## [0.4.0] - 2026-01-25
 
 ### Added